+++ src/bizy_server/error_handler.py
@@ -1,0 +1,3 @@
+class ErrorHandler:
+    def handle_error(self, error_code, message=None, data=None):
+        return {"code": error_code, "message": message or "", "data": data or {}}
